refactor: move scroll sidebar script logic to navbar component
